+++ ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: yizhang <zhaozicen951230@gmail.com>          +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/11/04 18:24:30 by yizhang       #+#    #+#                 */
-/*   Updated: 2023/03/12 19:57:06 by yizhang       ########   odam.nl         */
+/*   Updated: 2023/03/13 08:41:16 by yizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ void	ft_lstiter(t_list *lst, void (*f)(void *));
 
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
+char	*ft_strnstr(const char *haystack, const char *needle, int len);
 char	*ft_strdup(const char *s1);
 char	*strdup(const char *s1);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
